fix : minor fixes
